+++ monkey/monkey_island/cc/ui/src/components/configuration-components/InternalConfig.js
@@ -11,7 +11,6 @@ const sectionOrder = [
   'dropper',
   'classes',
   'general',
-  'kill_file',
   'testing'
 ];
 const initialSection = sectionOrder[0];